backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
stripe==7.8.0
email-validator==2.1.0
werkzeug==3.0.1
imagekitio==3.2.0
